requirements.txt
pandas==1.1.4
numpy==1.19.3
tqdm==4.51.0
nltk==3.5
beautifulsoup4==4.9.3
requests
requests_oauthlib==1.3.0
git+https://github.com/mobolic/facebook-sdk.git#egg=facebook-sdk
Mastodon.py==1.5.1
python-dotenv
python-dateutil==2.8.1
instabot
